+++ crawler/src/UrlFrontier.cpp
@@ -39,14 +39,41 @@ constexpr unsigned int URLHighScoreCutoff = 90;        // Score >= 90 is "high"
 constexpr unsigned int URLHighScoreQueuePercent = 90;  // Take from "high" scoring urls 90% of the time
 
 constexpr size_t MaxFreshURLsBatch = 50000;
+constexpr unsigned int FullGrowthRate = 10000;
+
+bool RandomAdmit(unsigned int bp) {
+    return rand() % FullGrowthRate < bp;
+}
+
+template<typename T>
+void SampleVectorInPlace(std::vector<T>& vec, unsigned int growthRateBp) {
+    if (growthRateBp >= FullGrowthRate) {
+        return;
+    }
+
+    size_t writeIdx = 0;
+    for (size_t readIdx = 0; readIdx < vec.size(); readIdx++) {
+        if (RandomAdmit(growthRateBp)) {
+            if (writeIdx != readIdx) {
+                vec[writeIdx] = std::move(vec[readIdx]);
+            }
+            writeIdx++;
+        }
+    }
+
+    vec.erase(vec.begin() + writeIdx, vec.end());
+}
+
 
 }  //  namespace
 
 UrlFrontier::UrlFrontier(HostRateLimiter* limiter,
                          const std::string& frontierDirectory,
+                         unsigned int growthRateBp,
                          size_t concurrentRobotsRequests,
                          size_t robotsCacheSize)
     : urlQueue_(frontierDirectory, URLHighScoreCutoff, URLHighScoreQueuePercent),
+      growthRateBp_(growthRateBp),
       robotRulesCache_(concurrentRobotsRequests, limiter, robotsCacheSize),
       delayCache_(robotsCacheSize) {}
 
@@ -103,11 +130,6 @@ void UrlFrontier::RobotsRequestsThread(ThreadSync& sync) {
 }
 
 void UrlFrontier::FreshURLsThread(ThreadSync& sync) {
-    {
-        core::LockGuard lock(urlQueueMu_);
-        FrontierSize.Set(urlQueue_.TotalSize());
-        FrontierQueueSize.Set(urlQueue_.Size());
-    }
     {
         core::LockGuard lock(freshURLsMu_);
         FrontierFreshURLs.Set(freshURLs_.size());
@@ -245,14 +267,35 @@ void UrlFrontier::GetURLs(ThreadSync& sync, size_t max, std::vector<std::string>
     GetURLsFiltered(sync, max, out, [](std::string_view) { return true; }, atLeastOne);
 }
 
-void UrlFrontier::PushURL(std::string u) {
+void UrlFrontier::PushURL(std::string u, bool always) {
+    if (!always) {
+        if (growthRateBp_ == 0) {
+            return;
+        } else if (growthRateBp_ < FullGrowthRate) {
+            if (!RandomAdmit(growthRateBp_)) {
+                return;
+            }
+        }
+    }
+
     core::LockGuard lock(freshURLsMu_);
     freshURLs_.push_back(std::move(u));
     FrontierFreshURLs.Set(freshURLs_.size());
     freshURLsCv_.Signal();
 }
 
-void UrlFrontier::PushURLs(std::vector<std::string>& urls) {
+void UrlFrontier::PushURLs(std::vector<std::string>& urls, bool always) {
+    if (!always) {
+        if (growthRateBp_ == 0) {
+            return;
+        }
+        SampleVectorInPlace(urls, growthRateBp_);
+    }
+
+    if (urls.empty()) {
+        return;
+    }
+
     core::LockGuard lock(freshURLsMu_);
     for (auto& url : urls) {
         freshURLs_.push_back(std::move(url));
@@ -266,7 +309,7 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
     long end;
 
     // 0. Wait for fresh URLs
-    std::vector<std::string> urls;
+    std::deque<std::string> urls;
     {
         core::LockGuard freshLock(freshURLsMu_);
         freshURLsCv_.Wait(freshLock, [&]() { return !freshURLs_.empty() || sync.ShouldSynchronize(); });
@@ -274,21 +317,23 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
             return;
         }
 
-        start = MonotonicTimeMs();
+        start = MonotonicTimeUs();
 
         ProcessFreshURLsCounter.Inc();
         if (freshURLs_.size() > MaxFreshURLsBatch) {
-            urls = std::vector<std::string>{freshURLs_.begin(), freshURLs_.begin() + MaxFreshURLsBatch};
-            freshURLs_.erase(freshURLs_.begin(), freshURLs_.begin() + MaxFreshURLsBatch);
+            // Only take first MaxFreshURLsBatch count
+            for (size_t i = 0; i < MaxFreshURLsBatch; ++i) {
+                urls.push_back(std::move(freshURLs_.front()));
+                freshURLs_.pop_front();
+            }
             FrontierFreshURLs.Set(freshURLs_.size());
         } else {
-            urls = std::move(freshURLs_);
-            freshURLs_.clear();
+            urls.swap(freshURLs_);
             FrontierFreshURLs.Zero();
         }
 
-        end = MonotonicTimeMs();
-        FreshURLsStepMove.Observe(static_cast<double>(end - start) / 1000.0);
+        end = MonotonicTimeUs();
+        FreshURLsStepMove.Observe(static_cast<double>(end - start) / 1000000.0 / static_cast<double>(urls.size()));
     }
 
     SPDLOG_TRACE("starting processing of {} fresh urls", urls.size());
@@ -297,7 +342,7 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
     validURLs.reserve(urls.size());
 
     // 1. Validate and parse URLs
-    start = MonotonicTimeMs();
+    start = MonotonicTimeUs();
 
     for (const auto& url : urls) {
         if (!IsValidUrl(url)) {
@@ -310,8 +355,8 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
         validURLs.push_back(std::move(*parsed));
     }
 
-    end = MonotonicTimeMs();
-    FreshURLsStepValidate.Observe(static_cast<double>(end - start) / 1000.0);
+    end = MonotonicTimeUs();
+    FreshURLsStepValidate.Observe(static_cast<double>(end - start) / 1000000.0 / static_cast<double>(urls.size()));
 
     if (validURLs.empty()) {
         SPDLOG_TRACE("finished processing of fresh urls: no valid urls");
@@ -319,23 +364,28 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
     }
 
     // 2. Discard already seen URLs
-    start = MonotonicTimeMs();
+    start = MonotonicTimeUs();
     std::vector<http::URL*> newURLs;
     newURLs.reserve(validURLs.size());
     {
         auto seen = std::set<std::string_view>{};
         core::LockGuard queueLock(urlQueueMu_);
         for (auto& url : validURLs) {
-            if (seen.contains(url.url) || urlQueue_.Seen(url.url)) {
+            if (seen.contains(url.url)) {
+                continue;
+            } else if (urlQueue_.Seen(url.url)) {
+                DuplicateURLCounter.Inc();
                 continue;
             }
             seen.insert(url.url);
             newURLs.push_back(&url);
+            NewURLCounter.Inc();
         }
     }
 
-    end = MonotonicTimeMs();
-    FreshURLsStepDeduplicate.Observe(static_cast<double>(end - start) / 1000.0);
+    end = MonotonicTimeUs();
+    FreshURLsStepDeduplicate.Observe(static_cast<double>(end - start) / 1000000.0 /
+                                     static_cast<double>(validURLs.size()));
 
     if (newURLs.empty()) {
         SPDLOG_TRACE("finished processing of fresh urls: no new urls");
@@ -354,7 +404,7 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
 
     // 4. Look up robots.txt ruleset (if in memory)
     assert(newURLs.size() == canonicalHosts.size());
-    start = MonotonicTimeMs();
+    start = MonotonicTimeUs();
     std::vector<RobotsLookupResult> robotResults;
     robotResults.reserve(canonicalHosts.size());
     {
@@ -371,12 +421,13 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
         }
     }
 
-    end = MonotonicTimeMs();
-    FreshURLsStepLookUpRobots.Observe(static_cast<double>(end - start) / 1000.0);
+    end = MonotonicTimeUs();
+    FreshURLsStepLookUpRobots.Observe(static_cast<double>(end - start) / 1000000.0 /
+                                      static_cast<double>(canonicalHosts.size()));
 
     // 5. Enqueue URLs that aren't ready to waiting list and discard disallowed URLs
     assert(newURLs.size() == robotResults.size());
-    start = MonotonicTimeMs();
+    start = MonotonicTimeUs();
     std::vector<http::URL*> pushURLs;
     pushURLs.reserve(newURLs.size());
     {
@@ -408,8 +459,8 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
         WaitingRobotsURLs.Set(urlsWaitingForRobotsCount_);
     }
 
-    end = MonotonicTimeMs();
-    FreshURLsStepEnqueue.Observe(static_cast<double>(end - start) / 1000.0);
+    end = MonotonicTimeUs();
+    FreshURLsStepEnqueue.Observe(static_cast<double>(end - start) / 1000000.0 / static_cast<double>(newURLs.size()));
 
     if (pushURLs.empty()) {
         SPDLOG_TRACE("finished processing of fresh urls: no ready urls, {} awaiting robots.txt", newURLs.size());
@@ -417,7 +468,7 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
     }
 
     // 6. Push all allowed, ready to fetch URLs onto frontier
-    start = MonotonicTimeMs();
+    start = MonotonicTimeUs();
     {
         core::LockGuard queueLock(urlQueueMu_);
         for (auto* url : pushURLs) {
@@ -428,8 +479,8 @@ void UrlFrontier::ProcessFreshURLs(ThreadSync& sync) {
     }
     urlQueueCv_.Broadcast();
 
-    end = MonotonicTimeMs();
-    FreshURLsStepPush.Observe(static_cast<double>(end - start) / 1000.0);
+    end = MonotonicTimeUs();
+    FreshURLsStepPush.Observe(static_cast<double>(end - start) / 1000000.0 / static_cast<double>(pushURLs.size()));
 
     SPDLOG_TRACE("finished processing of fresh urls: {} urls pushed, {} awaiting robots.txt",
                  pushURLs.size(),